Move class Encampment
